# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: NovaOS API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /workspaces:
    get:
      operationId: listWorkspaces
      tags: [collaboration]
      summary: List user's workspaces (owned + member)
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceList"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createWorkspace
      tags: [collaboration]
      summary: Create a new workspace
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                description:
                  type: string
                color:
                  type: string
                displayName:
                  type: string
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /workspaces/{workspaceId}:
    get:
      operationId: getWorkspace
      tags: [collaboration]
      summary: Get workspace detail with members
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceDetail"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateWorkspace
      tags: [collaboration]
      summary: Update workspace (owner only)
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
                color:
                  type: string
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Workspace"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteWorkspace
      tags: [collaboration]
      summary: Delete workspace (owner only)
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /workspaces/{workspaceId}/members:
    get:
      operationId: listWorkspaceMembers
      tags: [collaboration]
      summary: List members of a workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkspaceMember"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /workspaces/{workspaceId}/members/{memberId}:
    delete:
      operationId: removeWorkspaceMember
      tags: [collaboration]
      summary: Remove member or leave workspace
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /workspaces/{workspaceId}/invites:
    get:
      operationId: listWorkspaceInvites
      tags: [collaboration]
      summary: List pending invites (owner only)
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkspaceInvite"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: sendWorkspaceInvite
      tags: [collaboration]
      summary: Send an invite by email (owner only)
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email:
                  type: string
                role:
                  type: string
                  enum: [editor, viewer]
                inviterDisplayName:
                  type: string
      responses:
        "201":
          description: Invite sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceInvite"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /workspaces/{workspaceId}/invites/{inviteId}:
    delete:
      operationId: revokeWorkspaceInvite
      tags: [collaboration]
      summary: Revoke a pending invite (owner only)
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: inviteId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Revoked
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /workspaces/invites/accept:
    post:
      operationId: acceptWorkspaceInvite
      tags: [collaboration]
      summary: Accept an invite by token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token]
              properties:
                token:
                  type: string
                displayName:
                  type: string
      responses:
        "200":
          description: Accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  workspace:
                    $ref: "#/components/schemas/Workspace"
                  role:
                    type: string
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Already a member or invite already used
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /workspaces/{workspaceId}/activity:
    get:
      operationId: getWorkspaceActivity
      tags: [collaboration]
      summary: Get recent workspace activity feed
      parameters:
        - name: workspaceId
          in: path
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/WorkspaceActivity"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Phase 11 — Cloud Editor ───────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Workspace:
      type: object
      required: [id, ownerUserId, name, description, color, createdAt, updatedAt]
      properties:
        id:
          type: string
        ownerUserId:
          type: string
        name:
          type: string
        description:
          type: string
        color:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    WorkspaceMember:
      type: object
      required: [id, workspaceId, userId, displayName, role, joinedAt]
      properties:
        id:
          type: integer
        workspaceId:
          type: string
        userId:
          type: string
        displayName:
          type: string
        role:
          type: string
          enum: [owner, editor, viewer]
        joinedAt:
          type: string
          format: date-time
    WorkspaceDetail:
      allOf:
        - $ref: "#/components/schemas/Workspace"
        - type: object
          required: [members]
          properties:
            members:
              type: array
              items:
                $ref: "#/components/schemas/WorkspaceMember"
    WorkspaceInvite:
      type: object
      required: [id, workspaceId, inviterUserId, inviteeEmail, token, role, status, createdAt, expiresAt]
      properties:
        id:
          type: integer
        workspaceId:
          type: string
        inviterUserId:
          type: string
        inviteeEmail:
          type: string
        token:
          type: string
        role:
          type: string
          enum: [editor, viewer]
        status:
          type: string
          enum: [pending, accepted, revoked]
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
    WorkspaceActivity:
      type: object
      required: [id, workspaceId, actorUserId, actorDisplayName, action, createdAt]
      properties:
        id:
          type: integer
        workspaceId:
          type: string
        actorUserId:
          type: string
        actorDisplayName:
          type: string
        action:
          type: string
        contextJson:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
    WorkspaceList:
      type: object
      required: [owned, member]
      properties:
        owned:
          type: array
          items:
            $ref: "#/components/schemas/Workspace"
        member:
          type: array
          items:
            $ref: "#/components/schemas/Workspace"
...