# lib/api-zod/src/generated/types/workspace.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * NovaOS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Workspace {
  id: string;
  ownerUserId: string;
  name: string;
  description: string;
  color: string;
  createdAt: Date;
  updatedAt: Date;
}