# lib/api-zod/src/generated/types/workspaceActivity.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * NovaOS API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WorkspaceActivity {
  id: number;
  workspaceId: string;
  actorUserId: string;
  actorDisplayName: string;
  action: string;
  contextJson?: string | null;
  createdAt: Date;
}